+++ src/apps/DDBMuncher.js
@@ -14,7 +14,7 @@ import AdventureMunch from "../muncher/adventure/AdventureMunch.js";
 import ThirdPartyMunch from "../muncher/adventure/ThirdPartyMunch.js";
 import MuncherSettings from "../lib/MuncherSettings.js";
 import DDBMacros from "../effects/DDBMacros.js";
-import { importCacheLoad } from "../lib/DDBTemplateStrings.js";
+import { importCacheLoad } from "../lib/DDBReferenceLinker.js";
 import { updateWorldMonsters, resetCompendiumActorImages } from "../muncher/tools.js";
 import { parseBackgrounds } from "../muncher/backgrounds.js";
 import { parseTransports } from "../muncher/vehicles.js";