+++ src/muncher/backgrounds/backgrounds.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-await-in-loop */
 import logger from "../../logger.js";
 import { generateBackground } from "../../parser/character/bio.js";
-// import { parseTags } from "../../lib/DDBTemplateStrings.js";
+// import { parseTags } from "../../lib/DDBReferenceLinker.js";
 import DDBHelper from "../../lib/DDBHelper.js";
 // import { generateTable } from "../table.js";
 import DDBItemImporter from "../../lib/DDBItemImporter.js";