+++ src/muncher/classes/shared.js
@@ -4,7 +4,7 @@ import FileHelper from "../../lib/FileHelper.js";
 import logger from "../../logger.js";
 import DICTIONARY from "../../dictionary.js";
 import { generateTable } from "../table.js";
-import { parseTags } from "../../lib/DDBTemplateStrings.js";
+import { parseTags } from "../../lib/DDBReferenceLinker.js";
 import utils from "../../lib/utils.js";
 import SETTINGS from "../../settings.js";