+++ src/parser/item/common.js
@@ -1,5 +1,5 @@
 import DICTIONARY from "../../dictionary.js";
-import { parseTags } from "../../lib/DDBTemplateStrings.js";
+import { parseTags } from "../../lib/DDBReferenceLinker.js";
 
 export function getDescription(data) {
   const chatSnippet = data.definition.snippet ? data.definition.snippet : "";