+++ src/parser/vehicle/components.js
@@ -1,7 +1,7 @@
 import { newComponent } from "./templates/component.js";
 import DICTIONARY from "../../dictionary.js";
 import logger from "../../logger.js";
-import { parseTags } from "../../lib/DDBTemplateStrings.js";
+import { parseTags } from "../../lib/DDBReferenceLinker.js";
 
 const TYPE_MAPPING = {
   hull: "equipment",