+++ src/parser/vehicle/vehicle.js
@@ -10,7 +10,7 @@ import { getCapacity } from './capacity.js';
 import { FLIGHT_IDS, getMovement } from './movement.js';
 import { processComponents } from './components.js';
 import { ACTION_THRESHOLDS } from './threshold.js';
-import { parseTags } from '../../lib/DDBTemplateStrings.js';
+import { parseTags } from '../../lib/DDBReferenceLinker.js';
 
 // eslint-disable-next-line complexity
 async function parseVehicle(ddb, extra = {}) {